requirements.txt
pillow
colorama

